add RuleTypesDropdown component
